requirements.txt
google-generativeai
Pillow
numpy
torch
